requirements.txt
mysql-connector-python==8.2.0
Flask==3.0.0
python-dotenv==1.0.0
